html/semantics/embedded-content-0/media-elements/canPlayType_bogus_type.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.canPlayType() - bogus format</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-navigator-canplaytype">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").canPlayType("application/marks-fantasmagorical-format"),
  "",
  "audioElement.canPlayType() should return the empty string on a bogus format");
}, "audio.canPlayType() - bogus format");

test(function() {
 assert_equals(
  document.getElementById("v").canPlayType("application/marks-fantasmagorical-format"),
  "",
  "videoElement.canPlayType() should return the empty string on a bogus format");
}, "video.canPlayType() - bogus format");
  </script>
 </body>
</html>
